+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/repository/domain/channelconfig/internal/ChannelConfigEntity.kt
@@ -33,6 +33,7 @@ internal data class ChannelConfigInnerEntity(
     val name: String,
     val isTypingEvents: Boolean,
     val isReadEvents: Boolean,
+    val deliveryEventsEnabled: Boolean,
     val isConnectEvents: Boolean,
     val isSearch: Boolean,
     val isReactionsEnabled: Boolean,