+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/offline/Mother.kt
@@ -36,8 +36,12 @@ import io.getstream.chat.android.offline.repository.domain.message.internal.Remi
 import io.getstream.chat.android.offline.repository.domain.queryChannels.internal.QueryChannelsEntity
 import io.getstream.chat.android.offline.repository.domain.reaction.internal.ReactionEntity
 import io.getstream.chat.android.offline.repository.domain.threads.internal.ThreadEntity
+import io.getstream.chat.android.offline.repository.domain.user.internal.DeliveryReceiptsEntity
 import io.getstream.chat.android.offline.repository.domain.user.internal.PrivacySettingsEntity
+import io.getstream.chat.android.offline.repository.domain.user.internal.ReadReceiptsEntity
+import io.getstream.chat.android.offline.repository.domain.user.internal.TypingIndicatorsEntity
 import io.getstream.chat.android.offline.repository.domain.user.internal.UserEntity
+import io.getstream.chat.android.offline.repository.domain.user.internal.UserMuteEntity
 import io.getstream.chat.android.randomBoolean
 import io.getstream.chat.android.randomCID
 import io.getstream.chat.android.randomDate
@@ -45,6 +49,7 @@ import io.getstream.chat.android.randomDateOrNull
 import io.getstream.chat.android.randomDouble
 import io.getstream.chat.android.randomInt
 import io.getstream.chat.android.randomString
+import io.getstream.chat.android.randomStringOrNull
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.asExecutor
 import java.util.Date
@@ -56,13 +61,13 @@ internal fun randomUserEntity(
     name: String = randomString(),
     image: String = randomString(),
     role: String = randomString(),
-    createdAt: Date? = null,
-    updatedAt: Date? = null,
-    lastActive: Date? = null,
+    createdAt: Date? = randomDateOrNull(),
+    updatedAt: Date? = randomDateOrNull(),
+    lastActive: Date? = randomDateOrNull(),
     invisible: Boolean = randomBoolean(),
     privacySettings: PrivacySettingsEntity? = null,
     banned: Boolean = randomBoolean(),
-    mutes: List<String> = emptyList(),
+    mutes: List<UserMuteEntity> = emptyList(),
     teams: List<String> = emptyList(),
     teamsRole: Map<String, String> = emptyMap(),
     extraData: Map<String, Any> = emptyMap(),
@@ -82,6 +87,30 @@ internal fun randomUserEntity(
     teams = teams,
     teamsRole = teamsRole,
     extraData = extraData,
+)
+
+internal fun randomPrivacySettingsEntity(
+    typingIndicators: TypingIndicatorsEntity = TypingIndicatorsEntity(enabled = randomBoolean()),
+    readReceipts: ReadReceiptsEntity = ReadReceiptsEntity(enabled = randomBoolean()),
+    deliveryReceipts: DeliveryReceiptsEntity = DeliveryReceiptsEntity(enabled = randomBoolean()),
+): PrivacySettingsEntity = PrivacySettingsEntity(
+    typingIndicators = typingIndicators,
+    readReceipts = readReceipts,
+    deliveryReceipts = deliveryReceipts,
+)
+
+internal fun randomUserMuteEntity(
+    userId: String? = randomStringOrNull(),
+    targetId: String? = randomStringOrNull(),
+    createdAt: Date = randomDate(),
+    updatedAt: Date = randomDate(),
+    expires: Date? = randomDateOrNull(),
+): UserMuteEntity = UserMuteEntity(
+    userId = userId,
+    targetId = targetId,
+    createdAt = createdAt,
+    updatedAt = updatedAt,
+    expires = expires,
 )
 
 internal fun randomMessageEntity(