+++ stream-chat-android-ui-common/src/main/kotlin/io/getstream/chat/android/ui/common/utils/extensions/Message.kt
@@ -17,10 +17,13 @@
 package io.getstream.chat.android.ui.common.utils.extensions
 
 import io.getstream.chat.android.client.ChatClient
+import io.getstream.chat.android.client.utils.message.isDeleted
+import io.getstream.chat.android.client.utils.message.isEphemeral
 import io.getstream.chat.android.core.internal.InternalStreamChatApi
 import io.getstream.chat.android.models.Message
 import io.getstream.chat.android.models.MessageModerationAction
 import io.getstream.chat.android.models.ModerationAction
+import io.getstream.chat.android.models.SyncStatus
 import io.getstream.chat.android.models.User
 
 /**
@@ -31,6 +34,10 @@ public fun Message.isMine(chatClient: ChatClient): Boolean = chatClient.clientSt
 
 @InternalStreamChatApi
 public fun Message.isMine(currentUser: User?): Boolean = currentUser?.id == user.id
+
+@InternalStreamChatApi
+public fun Message.shouldShowMessageStatusIndicator(): Boolean =
+    !isEphemeral() && !isDeleted() && syncStatus != SyncStatus.FAILED_PERMANENTLY
 
 /**
  * @return if the message failed at moderation or not.